backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
telethon
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings
openai
sendgrid
twilio
python-dotenv
alembic
pytest
httpx
